+++ src/Commands/CreateKeys.php
@@ -22,26 +22,32 @@ class CreateKeys extends Command
      *
      * @var string
      */
-    protected $description = 'Creates Sodium public and private key.';
+    protected $description = 'Creates public and private crypto keys.';
 
     /**
      * Execute the console command.
      *
      * @return void
      */
-    public function handle()
+    public function handle(): void
     {
         $keyPair = sodium_crypto_box_keypair();
         $publicKey = sodium_crypto_box_publickey($keyPair);
 
-        $this->info('Successfully generated Sodium key pair.');
+        $this->newLine();
 
-        $this->info(sprintf('This is your public key: %s', sodium_bin2hex($publicKey)));
-        $this->warn('Please send the public key to the admin, so it can be saved in the database!');
+        $this->info('Successfully generated crypto key pair.');
 
-        $this->output->newLine();
+        $this->newLine();
 
-        $this->info('Please write the key pair in your .env file!');
-        $this->warn(sprintf('%s=%s', app('protector')->getPrivateKeyName(), sodium_bin2hex($keyPair)));
+        $this->comment('Send the public key to server admin so that it can be stored.');
+        $this->comment('Add the private key to your .env file.');
+
+        $this->newLine();
+
+        $this->info(sprintf('# Protector Public Key: %s', sodium_bin2hex($publicKey)));
+        $this->info(sprintf('%s=%s', app('protector')->getPrivateKeyName(), sodium_bin2hex($keyPair)));
+
+        $this->newLine();
     }
 }