+++ src/Commands/CreateToken.php
@@ -31,33 +31,35 @@ class CreateToken extends Command
      *
      * @return void
      */
-    public function handle()
+    public function handle(): void
     {
         $publicKey = $this->option('publicKey');
         $user = config('auth.providers.users.model')::findOrFail($this->argument('userId'));
         $user->tokens()->whereAbilities('["protector:import"]')->delete();
-        $userInformation = sprintf('%s: %s (%s)', $user->id, $user->name, $user->email);
+
+        $this->newLine();
+
+        $this->warn(sprintf('Executing for User %s|%s (%s)', $user->id, $user->name, $user->email));
 
         if (!$user->protector_public_key && !$publicKey) {
-            $this->error(
-                'The user doesn\'t have a protector public key and none was specified. Please provide a public key for the user.'
-            );
-            return null;
+            $this->fail('The user doesn\'t have a protector public key and none was specified. Please provide a public key for the user.');
         }
 
         if ($publicKey) {
             $user->protector_public_key = $publicKey;
             $user->save();
 
-            $this->info(sprintf('Protector public key was set for user %s.', $userInformation));
-            $this->output->newLine();
+            $this->info('Protector public key was set.');
         }
+
+        $this->newLine();
 
         $token = $user->createToken('protector', ['protector:import']);
 
-        $this->warn(sprintf('Information for the user %s', $userInformation));
-        $this->info(sprintf('%s="%s"', app('protector')->getAuthTokenKeyName(), $token->plainTextToken));
         $this->warn('The quotation marks at the start and end of the token are necessary!');
+        $this->info(sprintf('%s="%s"', app('protector')->getAuthTokenKeyName(), $token->plainTextToken));
         $this->info(sprintf('%s=%s', app('protector')->getDumpEndpointUrlKeyName(), route('protectorDumpEndpointRoute')));
+
+        $this->newLine();
     }
 }