+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to `protector` will be documented in this file.
 ### General
 
 - Restructured the `protector.php` configuration file for better organization and clarity
+- Reformatted the output of the `protector:keys` and `protector:token` commands to easier spot relevant information
 
 ### Features